Add new run_branch module that contains the support code to manage the run branch feature.
This consists of a RunBranch struct that implements the AvroSchema trait and all the necessary facility to publish the data to sasquatch. In addition, since we want to be able to retrieve the information from the EFD, the struct also contains methods to retrieve it from the EFD using the lsst_efd_client library.